refactor(proposals): consolidate code + annotated-code into one block
Fold the standalone Code block's language switcher and collapse-to-N-lines
(with "Show all N lines") into AnnotatedCode, making it the single general
code block: filename header + copy + line numbers + language switcher +
collapse + OPTIONAL line-anchored annotations. Remove the now-redundant
`code` block type (Rust enum/registry/tests, TS registry/components/parity
fixtures) and delete CodeSnippetBlock.tsx. Display name for annotated-code
is now "Code".
