+++ ui/src/components/proposals/blocks/__fixtures__/canonicalProposal.mdx
@@ -68,12 +68,6 @@ This runs on the hot path — keep allocations out of the loop.
 - [ ] Docs updated — link the runbook
 </Checklist>
 
-<Code id="snippet" filename="src/util.ts" lang="ts">
-export function clamp(value: number, min: number, max: number) {
-  return Math.max(min, Math.min(max, value));
-}
-</Code>
-
 <QuestionForm id="open-questions" title="Open Questions">
 Should we use Redis or Memcached for caching?
 </QuestionForm>
