+++ ui/src/components/proposals/blocks/index.ts
@@ -11,7 +11,6 @@ export { DiffBlock } from "./DiffBlock";
 export { CalloutBlock } from "./CalloutBlock";
 export { TableBlock } from "./TableBlock";
 export { ChecklistBlock } from "./ChecklistBlock";
-export { CodeSnippetBlock, CodeSurface } from "./CodeSnippetBlock";
 
 // Reusable line-anchored annotation engine (pure parsing + React rail UI). The
 // `annotated-code` block consumes it; `diff` and future blocks can adopt it.
